refactor(server): reorganize storage (no-op)
